==> data_prep_for_ml/2_Comparison with previous files.R <==
setwd('G:\\My Drive\\Research\\UMAA\\Data\\fresh_data')
library(data.table)
library(dplyr)
academic = fread('academic_cleaned.csv', select = 1L)
individual = fread('individual_info_cleaned.csv', select = 1L)
membership = fread('membership_cleaned.csv')
events = fread('events_cleaned.csv', select = 1L)
emails = fread('emails_cleaned.csv')
engagement = fread('engagement_cleaned.csv')

setwd('G:\\My Drive\\Research\\UMAA\\Data\\CAL Data\\Initial Data Files Cleaned')
academic_old = fread('academic_cleaned.csv', select = 1L)
individual_old = fread('individual_info_cleaned.csv', select = 1L)
membership_old = fread('membership_cleaned.csv')
events_old = fread('events_cleaned.csv', select = 1L)
emails_old = fread('emails_cleaned.csv')
engagement_old = fread('engagement_cleaned.csv')

year_level_mem <- engagement %>% group_by(YEAR_FISCAL, ENGAGEMENT_LEVEL_LIFE) %>% summarise(tot_mem = n_distinct(ID_DEMO))
year_level_mem_old <- engagement_old %>% group_by( YEAR_FISCAL, ENGAGEMENT_LEVEL_LIFE) %>% summarise(tot_mem = n_distinct(ID_DEMO))
